backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
python-dateutil==2.8.2
django-cors-headers==4.3.1


